requirements.txt
discord.py
regex
requests